+++ src/sdl/utils.cpp
@@ -338,9 +338,9 @@ surface scale_surface_legacy(const surface &surf, int w, int h)
 				int loc;
 				for (loc=0; loc<4; loc++) {
 				  a = pix[loc] >> 24;
-				  r = pix[loc] >> 16;
-				  g = pix[loc] >> 8;
-				  b = pix[loc] >> 0;
+				  r = static_cast<uint8_t>(pix[loc] >> 16);
+				  g = static_cast<uint8_t>(pix[loc] >> 8);
+				  b = static_cast<uint8_t>(pix[loc] >> 0);
 				  if (a != 0) {
 				    avg_r += r;
 				    avg_g += g;
@@ -366,9 +366,9 @@ surface scale_surface_legacy(const surface &surf, int w, int h)
 				rr = gg = bb = aa = 0;
 				for (loc=0; loc<4; loc++) {
 				  a = pix[loc] >> 24;
-				  r = pix[loc] >> 16;
-				  g = pix[loc] >> 8;
-				  b = pix[loc] >> 0;
+				  r = static_cast<uint8_t>(pix[loc] >> 16);
+				  g = static_cast<uint8_t>(pix[loc] >> 8);
+				  b = static_cast<uint8_t>(pix[loc] >> 0);
 				  if (a == 0) {
 				    r = static_cast<uint8_t>(avg_r);
 				    g = static_cast<uint8_t>(avg_g);
@@ -379,10 +379,10 @@ surface scale_surface_legacy(const surface &surf, int w, int h)
 				  bb += b * bilin[loc];
 				  aa += a * bilin[loc];
 				}
-				r = rr >> 16;
-				g = gg >> 16;
-				b = bb >> 16;
-				a = aa >> 16;
+				r = static_cast<uint8_t>(rr >> 16);
+				g = static_cast<uint8_t>(gg >> 16);
+				b = static_cast<uint8_t>(bb >> 16);
+				a = static_cast<uint8_t>(aa >> 16);
 				*dst_word = (a << 24) + (r << 16) + (g << 8) + b;
 			}
 		}
